feat(logger): improve default log formatter ordering
This resolves #1568. The problem raised in that issue is that the log
formatter outputs log lines with keys written in alphabetical order. In
practice, that means they start off with the cold start flag, the
`function_arn`, `function_name`, `function_memory_size`, and a whole
bunch of other stuff that is not usually of interest. The actual log
level and message are buried and need log lines expanded to be seen.
Even when expanded, they're hard to visually spot at a glance. When
viewing a log group in Cloudwatch or some other UI that truncates log
lines, you effectively see no useful information at all (see screenshot
in #1568 for example).

This change addresses the problem by surfacing log properties in order
of usefulness. This is a subjective decision, and opinions may vary, but
the order submitted in this PR should be preferable for the vast
majority of users. The "at-a-glance" view of logs should now be
actionable and help with log-diving and debugging. The actual
information recorded is unchanged and log insight queries should still
work as before.

The serialization order is well-defined in modern ECMAScript engines:

Serialization From the MDN docs on `JSON.stringify`:

> Only enumerable own properties are visited. This means `Map`, `Set`,
> etc. will become `"{}"`. You can use the replacer parameter to
> serialize them to something more useful. Properties are visited using
> the same algorithm as `Object.keys()`, which has a well-defined order
> and is stable across implementations. For example, `JSON.stringify` on
> the same object will always produce the same string, and
> `JSON.parse(JSON.stringify(obj))` would produce an object with the
> same key ordering as the original (assuming the object is completely
> JSON-serializable).
>
> Source: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/JSON/stringify#description

And the note about what that order is:

> The traversal order, as of modern ECMAScript specification, is
> well-defined and consistent across implementations. Within each
> component of the prototype chain, all non-negative integer keys (those
> that can be array indices) will be traversed first in ascending order
> by value, then other string keys in ascending chronological order of
> property creation.
>
> Source: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Statements/for...in